Moves the tests of program.rs into a separate file.
