+++ tests/vm.rs
@@ -1,7 +1,8 @@
 use solana_sbpf::{
     elf::Executable,
-    program::BuiltinProgram,
-    vm::{RuntimeEnvironmentSlot, TestContextObject},
+    program::{BuiltinFunction, BuiltinProgram, FunctionRegistry},
+    syscalls,
+    vm::{Config, RuntimeEnvironmentSlot, TestContextObject},
 };
 use std::{fs::File, io::Read, sync::Arc};
 use test_utils::create_vm;
@@ -48,4 +49,31 @@ fn test_runtime_environment_slots() {
     check_slot!(env, registers, Registers);
     check_slot!(env, program_result, ProgramResult);
     check_slot!(env, memory_mapping, MemoryMapping);
+}
+
+#[test]
+fn test_builtin_program_eq() {
+    let mut function_registry_a = FunctionRegistry::<BuiltinFunction<TestContextObject>>::default();
+    function_registry_a
+        .register_function_hashed(*b"log", syscalls::SyscallString::vm)
+        .unwrap();
+    function_registry_a
+        .register_function_hashed(*b"log_64", syscalls::SyscallU64::vm)
+        .unwrap();
+    let mut function_registry_b = FunctionRegistry::<BuiltinFunction<TestContextObject>>::default();
+    function_registry_b
+        .register_function_hashed(*b"log_64", syscalls::SyscallU64::vm)
+        .unwrap();
+    function_registry_b
+        .register_function_hashed(*b"log", syscalls::SyscallString::vm)
+        .unwrap();
+    let mut function_registry_c = FunctionRegistry::<BuiltinFunction<TestContextObject>>::default();
+    function_registry_c
+        .register_function_hashed(*b"log_64", syscalls::SyscallU64::vm)
+        .unwrap();
+    let builtin_program_a = BuiltinProgram::new_loader(Config::default(), function_registry_a);
+    let builtin_program_b = BuiltinProgram::new_loader(Config::default(), function_registry_b);
+    assert_eq!(builtin_program_a, builtin_program_b);
+    let builtin_program_c = BuiltinProgram::new_loader(Config::default(), function_registry_c);
+    assert_ne!(builtin_program_a, builtin_program_c);
 }
